requirements.txt
Django==1.11.8
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
image==1.5.24
Pillow==5.1.0
PyJWT==1.6.4
pytz==2018.4